requirements.txt
beautifulsoup4
isort
jupyter
jupytext
nbconvert
pylint-pytest
rpy2
